CreateACTActivity's firebase functionality refactored into the createACT method in ACTObject
