+++ app/src/main/java/com/example/mariokarttournamentorganizer/CreateACTActivity.java
@@ -7,30 +7,11 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.Timestamp;
-import com.google.firebase.firestore.AggregateQuery;
-import com.google.firebase.firestore.AggregateQuerySnapshot;
-import com.google.firebase.firestore.AggregateSource;
-import com.google.firebase.firestore.DocumentReference;
-import com.google.firebase.firestore.DocumentSnapshot;
-import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.Query;
-import com.google.firebase.firestore.FirebaseFirestoreException;
-import com.google.firebase.firestore.EventListener;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.Arrays;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
 
 public class CreateACTActivity extends AppCompatActivity {
     private Button createButton;
@@ -40,30 +21,16 @@ public class CreateACTActivity extends AppCompatActivity {
     private String timeStr;
     private EditText location;
     private String locationStr;
-
-    //for firebase
-    public static final String ID_FIELD = "ID";
-    public static final String ADMIN_FIELD = "adminID";
-    public static final String DATE_FIELD = "date";
-    public static final String TIME_FIELD = "time";
-    public static final String LOCATION_FIELD = "location";
-    public static final String PLAYERS_FIELD = "players";
     public static final String TAG = "creatingACT";
-    public static final String COMPLETED_FIELD = "completed";
-    public static final String RESULT_FIELD = "result";
-    public static final String TIMESTAMP_FIELD = "Date/Time";
-    private long count;
-    private String docPath;
-
-    private FirebaseFirestore myFireStore = FirebaseFirestore.getInstance();
-
+    private ACTObject actObj;
 
     @Override
-    protected void
-
-    onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.createact);
+
+        // making ACTObject
+        actObj = new ACTObject();
 
         //getting the username from the homescreen, to use for the adminID and the players list
         Intent loginInfoIntent = getIntent();
@@ -75,26 +42,6 @@ public class CreateACTActivity extends AppCompatActivity {
         time = (EditText) findViewById(R.id.editTextTime);
         location = (EditText) findViewById(R.id.editTextLocation);
 
-        //getting count of act_objects to generate new ID
-        Query query = myFireStore.collection("act_objects");
-        AggregateQuery countQuery = query.count();
-        countQuery.get(AggregateSource.SERVER).addOnCompleteListener(new OnCompleteListener<AggregateQuerySnapshot>() {
-            @Override
-            public void onComplete(@NonNull Task<AggregateQuerySnapshot> task) {
-                if (task.isSuccessful()) {
-                    // Count fetched successfully
-                    AggregateQuerySnapshot snapshot = task.getResult();
-                    count = snapshot.getCount();
-                    docPath = "act" + count;
-                    Log.d(TAG, "Count: " + count);
-                    Log.d(TAG, "Document name:" + docPath);
-                } else {
-                    Log.d(TAG, "Count failed: ", task.getException());
-                }
-            }
-        });
-
-
         //Set up create button
         createButton = (Button) findViewById(R.id.createButton);
         createButton.setOnClickListener(v->
@@ -105,37 +52,8 @@ public class CreateACTActivity extends AppCompatActivity {
 
             try {
                 if(!dateStr.isEmpty() && !timeStr.isEmpty() && !locationStr.isEmpty() && checkDate(dateStr)) {
-
-                    //creating a new document in firebase
-                    Map<String, Object> act1 = new HashMap<>();
-                    act1.put(ID_FIELD, count);
-                    act1.put(ADMIN_FIELD, usernameStr);
-                    act1.put(TIMESTAMP_FIELD, Timestamp.now());
-                    act1.put(COMPLETED_FIELD, false); //default
-                    act1.put(DATE_FIELD, dateStr);
-                    act1.put(TIME_FIELD, timeStr);
-                    act1.put(LOCATION_FIELD, locationStr);
-                    act1.put(PLAYERS_FIELD, Arrays.asList(usernameStr));
-                    act1.put(RESULT_FIELD, null);
-
-                    myFireStore.collection("act_objects").document(docPath)
-                            .set(act1)
-                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                @Override
-                                public void onSuccess(Void aVoid) {
-                                    Log.d(TAG, "DocumentSnapshot successfully written!");
-                                }
-                            })
-                            .addOnFailureListener(new OnFailureListener() {
-                                @Override
-                                public void onFailure(@NonNull Exception e) {
-                                    Log.w(TAG, "Error writing document", e);
-                                }
-                            });
-
-                    //Go back to home screen after created
-                    Intent homeScreen = new Intent(this, homeScreenActivity.class);
-                    startActivity(homeScreen);
+                    actObj.createACT(usernameStr, dateStr, timeStr, locationStr);
+                    finish();
                 }
                 else {
                     Toast.makeText(CreateACTActivity.this,
